+++ CHANGELOG.md
@@ -4,11 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-# [Unreleased]
+# [5.3.0] - 2023-06-10
 # Added
 - Added UMM option to change spells known per level.
 - Added UMM option to disable spellbooks for each Isekai archetype.
 - Added UMM option to change Isekai Protagonist to have the Shifter appearance.
+# Fixed
+- Fixed Shifter Legacy aspects minor forms not scaling.
 
 # [5.2.3] - 2023-05-30
 # Fixed